requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
pynacl==1.5.0
python-dotenv==1.0.0
jinja2==3.1.2
python-multipart==0.0.6
